querier/worker: improve code-style in tests
